Added accounting/audit and user deposits
The system now allows for deposits to unit's balance, fees are paid automatically via a cron job. Every transaction is logged, payments include generated pdf receipts. Users can upload and view documents for audit purposes.

+++ backend/src/main/java/com/smartentrance/backend/dto/poll/VoteCastRequest.java
@@ -6,5 +6,5 @@ public record VoteCastRequest(
         @NotNull(message = "You must select an option to vote")
         Integer optionId,
         @NotNull(message = "You must specify the unit which is voting.")
-        Integer unitId
+        Long unitId
 ) {}
